+++ src/App.vue
@@ -4,54 +4,105 @@
     <div class="container">
       <div class="card has-background-dark">
         <header class="card-header">
-          <p class="card-header-title has-text-white">Speedtest Ergebnisse</p>
+          <p class="card-header-title has-text-white">
+            Speedtest Ergebnisse
+          </p>
           <p class="has-text-white item-center">
-            <button :disabled="page <= 1" class="button is-small is-white" @click="setPage(page-1)">
+            <span v-if="data !== null" style="margin-lef: 2em;"
+              >{{ data.length }} Einträge gesamt</span
+            >
+          </p>
+          <p class="has-text-white item-center">
+            <button
+              :disabled="page <= 1"
+              class="button is-small is-white"
+              @click="setPage(page - 1)"
+            >
               &lt;
             </button>
           </p>
-          <p class="has-text-white" style="align-items: center; display: flex; flex-grow: 0;">
+          <p
+            class="has-text-white"
+            style="align-items: center; display: flex; flex-grow: 0;"
+          >
             <span class="tag is-light has-text-centered" style="width: 4em;">
               {{ page }}/{{ maxPage }}
             </span>
           </p>
           <p class="has-text-white item-center">
-            <button :disabled="page >= maxPage" class="button is-small is-white" @click="setPage(page+1)">
+            <button
+              :disabled="page >= maxPage"
+              class="button is-small is-white"
+              @click="setPage(page + 1)"
+            >
               &gt;
             </button>
           </p>
         </header>
         <div class="card-content" style="">
           <div class="content">
-            <div class="field has-addons is-pulled-right" style="margin-bottom: 0;">
+            <div
+              class="field has-addons is-pulled-right"
+              style="margin-bottom: 0;"
+            >
               <p class="control">
-                <a class="button is-small is-white" :class="{'is-active': objPerPage == 5}" @click="setObjects(5)">
+                <a
+                  class="button is-small is-white"
+                  :class="{ 'is-active': objPerPage == 5 }"
+                  @click="setObjects(5)"
+                >
                   <span>5</span>
                 </a>
               </p>
               <p class="control">
-                <a class="button is-small is-white" :class="{'is-active': objPerPage == 10}" @click="setObjects(10)">
+                <a
+                  class="button is-small is-white"
+                  :class="{ 'is-active': objPerPage == 10 }"
+                  @click="setObjects(10)"
+                >
                   <span>10</span>
                 </a>
               </p>
               <p class="control">
-                <a class="button is-small is-white" :class="{'is-active': objPerPage == 15}" @click="setObjects(15)">
+                <a
+                  class="button is-small is-white"
+                  :class="{ 'is-active': objPerPage == 15 }"
+                  @click="setObjects(15)"
+                >
                   <span>15</span>
                 </a>
               </p>
               <p class="control">
-                <a class="button is-small is-white" :class="{'is-active': objPerPage == 20}" @click="setObjects(20)">
+                <a
+                  class="button is-small is-white"
+                  :class="{ 'is-active': objPerPage == 20 }"
+                  @click="setObjects(20)"
+                >
                   <span>20</span>
                 </a>
               </p>
             </div>
+            <label class="checkbox is-pulled-right" style="margin-right: 1em;">
+              <input v-model="checked" type="checkbox" />
+              Nur Problemfälle anzeigen
+            </label>
             <div class="is-clearfix" />
-            <div v-if="data === null || data.length == 0" class="notification is-warning" style="margin-top: 1em;">
-              <span v-if="!error">Es stehen aktuell leider keine Daten zur Verfügung!<br>
-              Bitte vergewissern Sie sich, dass der Cronjob korrekt läuft.</span>
-              <span v-else>{{error}}</span>
+            <div
+              v-if="data === null || data.length == 0"
+              class="notification is-warning"
+              style="margin-top: 1em;"
+            >
+              <span v-if="!error"
+                >Es stehen aktuell leider keine Daten zur Verfügung!<br />
+                Bitte vergewissern Sie sich, dass der Cronjob korrekt
+                läuft.</span
+              >
+              <span v-else>{{ error }}</span>
             </div>
-            <table v-else class="table has-background-dark has-text-white is-narrow">
+            <table
+              v-else
+              class="table has-background-dark has-text-white is-narrow"
+            >
               <thead>
                 <tr>
                   <th>Datum</th>
@@ -63,23 +114,42 @@
               </thead>
               <tbody>
                 <tr
-                  v-for="(data, index) in data.slice((page-1)*objPerPage, (page*objPerPage))"
-                  :key="index" :style="(data.dl<minimalValues[0] || data.ul<minimalValues[1]) ? { 'background-color': '#463636'} : ''"
+                  v-for="(item, index) in computedData"
+                  :key="index"
+                  :style="
+                    item.dl < minimalValues[0] || item.ul < minimalValues[1]
+                      ? { 'background-color': '#463636' }
+                      : ''
+                  "
                 >
-                  <td data-label="Datum: ">{{data.date}}</td>
-                  <td data-label="Zeit: ">{{data.time}}</td>
-                  <td data-label="Ping: ">{{data.ping}}</td>
-                  <td data-label="DL: " :class="{'has-text-danger':(data.dl<minimalValues[0])}">{{data.dl}}</td>
-                  <td data-label="UL: " :class="{'has-text-danger':(data.ul<minimalValues[1])}">{{data.ul}}</td>
+                  <td data-label="Datum: ">{{ item.date }}</td>
+                  <td data-label="Zeit: ">{{ item.time }}</td>
+                  <td data-label="Ping: ">
+                    {{ parseFloat(item.ping).toFixed(2) }}
+                  </td>
+                  <td
+                    data-label="DL: "
+                    :class="{ 'has-text-danger': item.dl < minimalValues[0] }"
+                  >
+                    {{ parseFloat(item.dl).toFixed(2) }}
+                  </td>
+                  <td
+                    data-label="UL: "
+                    :class="{ 'has-text-danger': item.ul < minimalValues[1] }"
+                  >
+                    {{ parseFloat(item.ul).toFixed(2) }}
+                  </td>
                 </tr>
               </tbody>
             </table>
           </div>
         </div>
         <footer class="card-foot" style="padding: 1em;">
-          &copy; <span class="is-hidden-mobile">Copyright</span> {{ new Date().getFullYear() }} -
-          <a href="https://www.tschoepel.de/" target="_blank" rel="noopener">Tschoepel.de</a>.
-          Alle Rechte vorbehalten.
+          &copy; <span class="is-hidden-mobile">Copyright</span>
+          {{ new Date().getFullYear() }} -
+          <a href="https://www.tschoepel.de/" target="_blank" rel="noopener"
+            >Tschoepel.de</a
+          >. Alle Rechte vorbehalten.
         </footer>
       </div>
     </div>
@@ -94,7 +164,7 @@ export default {
   components: {
     // HelloWorld
   },
-  data () {
+  data() {
     return {
       loading: false,
       err: false,
@@ -102,15 +172,46 @@ export default {
       data: null,
       page: 1,
       objPerPage: 10,
-      minimalValues: [999, 999]
+      minimalValues: [999, 999],
+      checked: false
     }
   },
-  mounted () {
+  computed: {
+    url() {
+      return process.env.NODE_ENV === 'development'
+        ? 'http://localhost:3030'
+        : ''
+    },
+    maxPage: function() {
+      return this.data !== null
+        ? Math.ceil(this.data.length / this.objPerPage)
+        : null
+    },
+    computedData: function() {
+      if (!this.checked) {
+        return this.data.slice(
+          (this.page - 1) * this.objPerPage,
+          this.page * this.objPerPage
+        )
+      } else {
+        var temp = this.data.filter(item => {
+          return (
+            item.dl < this.minimalValues[0] || item.ul < this.minimalValues[1]
+          )
+        })
+        return temp.slice(
+          (this.page - 1) * this.objPerPage,
+          this.page * this.objPerPage
+        )
+      }
+    }
+  },
+  mounted() {
     this.get()
     this.getMin()
   },
   methods: {
-    get () {
+    get() {
       this.loading = true
       this.$http
         .get(this.url + '/api')
@@ -121,14 +222,19 @@ export default {
         })
         .catch(error => {
           console.error(error)
-          if (typeof error.response !== 'undefined' && typeof error.response.data !== 'undefined') {
+          if (
+            typeof error.response !== 'undefined' &&
+            typeof error.response.data !== 'undefined'
+          ) {
             error = error.response.data
           }
           this.err = true
         })
-        .then(() => { this.loading = false })
+        .then(() => {
+          this.loading = false
+        })
     },
-    getMin () {
+    getMin() {
       this.$http
         .get(this.url + '/min')
         .then(response => {
@@ -141,38 +247,31 @@ export default {
           this.error = error
         })
     },
-    setPage (page) {
+    setPage(page) {
       if (page < 1) page = 1
       if (page > this.maxPage) page = this.maxPage
       this.page = page
     },
-    setObjects (num) {
+    setObjects(num) {
       this.objPerPage = num
       if (this.page > this.maxPage) {
         this.page = this.maxPage
       }
-    }
-  },
-  computed: {
-    url () {
-      return (process.env.NODE_ENV === 'development') ? 'http://localhost:3030' : ''
-    },
-    maxPage: function () {
-      return (this.data !== null) ? Math.ceil(this.data.length / this.objPerPage) : null
     }
   }
 }
 </script>
 
 <style lang="scss">
-@import "~bulma";
-body, html {
+@import '~bulma';
+body,
+html {
   height: 100%;
   background-color: #1d2127;
   color: #808697;
 }
 .section {
-    padding: 1.5rem;
+  padding: 1.5rem;
 }
 .table th {
   color: inherit !important;
@@ -195,7 +294,7 @@ body, html {
 
 .card-header .button {
   // margin-top: 11px;
-  margin: 0.70rem;
+  margin: 0.7rem;
 }
 
 .card-foot {
@@ -203,10 +302,10 @@ body, html {
   border-top: 1px solid #1d2127;
 }
 a {
-  color: hsl(217, 71%, 68%)
+  color: hsl(217, 71%, 68%);
 }
 a:hover {
-  color: hsl(217, 71%, 78%)
+  color: hsl(217, 71%, 78%);
 }
 
 .item-center {
@@ -214,23 +313,28 @@ a:hover {
   display: flex;
   flex-grow: 0;
 }
-.content p:not(:last-child){
+.content p:not(:last-child) {
   margin-bottom: 0;
 }
 .table.is-striped tbody tr:not(.is-selected):nth-child(even) {
   background-color: hsla(0, 0%, 25%, 1);
 }
-.content table td, .content table th {
+.content table td,
+.content table th {
   border: 1px solid hsl(0, 0%, 18%);
   border-width: 1px 0px;
   padding: 0.5em 0.75em;
   vertical-align: top;
 }
 
-@media screen and (max-width:500px) {
+.checkbox:hover,
+.radio:hover {
+  color: white;
+}
 
+@media screen and (max-width: 500px) {
   thead {
-    display:none;
+    display: none;
   }
   table {
     margin-top: 1em;
@@ -245,15 +349,14 @@ a:hover {
   td {
     float: left;
     width: 100%;
-    padding:1em;
+    padding: 1em;
   }
 
   td::before {
-    content:attr(data-label);
+    content: attr(data-label);
     box-sizing: border-box;
     word-wrap: break-word;
     width: 20%;
   }
 }
-
 </style>